requirements.txt
openai>=1.30.0
PyGithub>=2.3.0
python-dotenv>=1.0.0
rich>=13.7.0
reportlab>=4.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
PyPDF2>=3.0.0
python-docx>=1.1.0
aiosqlite>=0.20.0
pyttsx3>=2.90
sounddevice>=0.4.6
scipy>=1.12.0
